dflash: pinned-HMMA GEMM dispatch in drafter forward — replaces 35 scalar GEMMs
Per spec amendment §6.1.A, all 35 projection matmuls (Q/K/V/O/Gate/Up/Down
× 5 layers) in dflash_drafter_forward_launch now dispatch through
ggml_cuda_mul_mat_f16_pinned (HMMA m16n8k16, byte-identity-by-construction
across batch composition). The scalar gemm_row_x_col_kernel body is
retired; cuBLAS HGEMM remains forbidden in this TU.

GEMM outputs are F32; downstream consumers (q_norm_rope, k_norm_rope,
cache_write_kv, silu_mul, residual_add) carry F32 in their input path
and fuse the F32→F16 cast at their natural store boundaries. Per-layer
data flow:

  F16 hidden_n → pinned Q/K/V (F32 outputs)
  F32 q_buf → q_norm_rope_kernel → F16 q_for_attn → attention_kernel
  F32 k_buf → k_norm_rope_kernel → F16 k_for_cache → cache_write_kv
  F32 v_buf → cache_write_kv (cast on store)
  F16 attn_out → pinned O (F32) → residual_add (F32+F16→F16) → F16 hidden
  F16 hidden_n → pinned Gate/Up (F32) → silu_mul_f32f32f16 → F16 act_buf
  F16 act_buf → pinned Down (F32) → residual_add (F32+F16→F16) → F16 hidden

Boot-time K%16 precondition check guards all four distinct K dimensions
(D_emb=5120, H_qDh=4096, H_kvDh=1024, intermediate=17408 — all ÷16 for
Qwen 3.6 27B drafter).

Also fixes test-dflash-drafter-lm-head.cpp which still referenced the
pre-recast BF16 lm_head signature (a F16 lm_head migration miss).

Gates green on production/2026-q2-next:
- test-dflash-drafter-lm-head: 5/5 byte-identical
- test-dflash-closure (DFLASH_TARGET_GGUF=lm_head-f16 target):
  8/8 prompts argmax 4/4 vs vLLM, cos ≥ 0.9998, NMSE ≤ 4e-4 per prompt
- test-dflash-np-invariance: 4/4 seeds × NP ∈ {1,2,4,8} byte-identical
- test-dflash-np-multislot: slot-0 byte-identical across NP ∈ {1,2,4,8}

+++ tests/dflash-speculative/test-dflash-drafter-lm-head.cpp
@@ -1,8 +1,8 @@
 // test-dflash-drafter-lm-head.cpp
 //
-// Byte-identity / NMSE unit test for dflash_drafter_lm_head — the BF16
-// GEMV against target's shared `output.weight`. Kernel output vs CPU
-// scalar fp32 reference across a (n_rows × V × D_emb × seed) sweep.
+// Byte-identity / NMSE unit test for dflash_drafter_lm_head — the F16
+// GEMV against target's shared `output.weight` (T1-recast). Kernel output
+// vs CPU scalar fp32 reference across a (n_rows × V × D_emb × seed) sweep.
 //
 // @witnesses: SharedEmbedAndLMHead
 //
@@ -18,7 +18,6 @@
 
 #include "ggml-cuda/dflash/dflash-drafter-lm-head.cuh"
 
-#include <cuda_bf16.h>
 #include <cuda_fp16.h>
 #include <cuda_runtime.h>
 
@@ -44,11 +43,11 @@ namespace {
 
 // CPU scalar reference. Per-cell serial fp32 dot product over D_emb.
 // Mirrors the kernel's accumulation order exactly: hidden lifted to
-// fp32 via __half2float, BF16 weight lifted to fp32, serial K-loop sum.
+// fp32 via __half2float, F16 weight lifted to fp32, serial K-loop sum.
 void lm_head_scalar_ref(
-    const __half        * hidden,       // [n_rows, D_emb]
-    const __nv_bfloat16 * lm_head_w,    // [V, D_emb]
-    float               * logits,       // [n_rows, V]
+    const __half * hidden,       // [n_rows, D_emb]
+    const __half * lm_head_w,    // [V, D_emb]
+    float        * logits,       // [n_rows, V]
     int n_rows, int D_emb, int V)
 {
     std::vector<float> hidden_f32(D_emb);
@@ -60,7 +59,7 @@ void lm_head_scalar_ref(
             float acc = 0.0f;
             for (int k = 0; k < D_emb; ++k) {
                 acc += hidden_f32[k] *
-                       __bfloat162float(lm_head_w[static_cast<std::size_t>(col) * D_emb + k]);
+                       __half2float(lm_head_w[static_cast<std::size_t>(col) * D_emb + k]);
             }
             logits[static_cast<std::size_t>(row) * V + col] = acc;
         }
@@ -78,9 +77,9 @@ int run_one(int n_rows, int D_emb, int V, uint32_t seed) {
     const std::size_t n_o = static_cast<std::size_t>(n_rows) * V;
 
     std::vector<__half> hidden_h(n_h);
-    std::vector<__nv_bfloat16> lm_head_w_h(n_w);
+    std::vector<__half> lm_head_w_h(n_w);
     for (auto & x : hidden_h) x = __float2half(dist(rng));
-    for (auto & x : lm_head_w_h) x = __float2bfloat16(dist(rng) * 0.05f);
+    for (auto & x : lm_head_w_h) x = __float2half(dist(rng) * 0.05f);
 
     // CPU reference
     std::vector<float> ref_logits(n_o);
@@ -89,14 +88,14 @@ int run_one(int n_rows, int D_emb, int V, uint32_t seed) {
         n_rows, D_emb, V);
 
     // GPU kernel
-    __half        * d_hidden = nullptr;
-    __nv_bfloat16 * d_lm    = nullptr;
-    float         * d_out   = nullptr;
+    __half * d_hidden = nullptr;
+    __half * d_lm     = nullptr;
+    float  * d_out    = nullptr;
     CUDA_CHECK(cudaMalloc(&d_hidden, n_h * sizeof(__half)));
-    CUDA_CHECK(cudaMalloc(&d_lm,     n_w * sizeof(__nv_bfloat16)));
+    CUDA_CHECK(cudaMalloc(&d_lm,     n_w * sizeof(__half)));
     CUDA_CHECK(cudaMalloc(&d_out,    n_o * sizeof(float)));
-    CUDA_CHECK(cudaMemcpy(d_hidden, hidden_h.data(), n_h * sizeof(__half), cudaMemcpyHostToDevice));
-    CUDA_CHECK(cudaMemcpy(d_lm,     lm_head_w_h.data(), n_w * sizeof(__nv_bfloat16), cudaMemcpyHostToDevice));
+    CUDA_CHECK(cudaMemcpy(d_hidden, hidden_h.data(),    n_h * sizeof(__half), cudaMemcpyHostToDevice));
+    CUDA_CHECK(cudaMemcpy(d_lm,     lm_head_w_h.data(), n_w * sizeof(__half), cudaMemcpyHostToDevice));
 
     dflash_drafter_lm_head_launch(
         d_hidden, d_lm, d_out, n_rows, D_emb, V, 0);
